Updates to calling default psk selection fn and enforcing a max no of psk identities
